backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector operations
pgvector==0.2.3
numpy==1.26.2

# Task queue
celery[redis]==5.3.4
flower==2.0.1
redis>=4.5.2,<5.0.0

# Web scraping
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
pyyaml==6.0.1
playwright==1.40.0
fake-useragent==1.5.1
robotexclusionrulesparser==1.6.2
praw==7.7.1

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1